% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_env.R
\docType{data}
\name{connections_rodbc}
\alias{connections_rodbc}
\title{Environment to store rodbc db connections}
\format{
An object of class \code{environment} of length 0.
}
\usage{
connections_rodbc
}
\description{
Environment to store rodbc db connections
}
\keyword{datasets}
