% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{get_db_connection}
\alias{get_db_connection}
\title{get_db_connection}
\usage{
get_db_connection(driver = NULL, server = NULL, db = NULL,
  port = NULL, user = NULL, password = NULL,
  db_config = config$db_config)
}
\arguments{
\item{driver}{driver}

\item{server}{server}

\item{db}{db}

\item{port}{port}

\item{user}{user}

\item{password}{password}

\item{db_config}{A list containing driver, server, db, port, user, password}
}
\description{
get_db_connection
}
