% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norsyss_download_diagnoses.R
\name{norsyss_download_aggregated_diagnoses}
\alias{norsyss_download_aggregated_diagnoses}
\title{Download NorSySS aggregated diagnoses}
\usage{
norsyss_download_aggregated_diagnoses(
  date_from = "2020-01-01",
  date_to = lubridate::today(),
  folder = getwd(),
  file_name = glue::glue("norsyss_{lubridate::today()}.txt"),
  diags = list(influensa = c("R80"), gastro = c("D11", "D70", "D73"), respiratory =
    c("R05", "R74", "R78", "R83"), respiratoryexternal = c("R05", "R74", "R78", "R83"),
    respiratoryinternal = c("R05", "R74", "R83"), lungebetennelse = c("R81"), bronkitt =
    c("R78"), skabb = c("S72"), emerg1 = c("R80"), emerg2 = c("R80"), emerg3 = c("R80"),
    emerg4 = c("R80"), emerg5 = c("R80")),
  overwrite_file = FALSE,
  ...
)
}
\arguments{
\item{date_from}{Start date for extracting data}

\item{date_to}{End date for extracting data}

\item{folder}{Folder the data file will end up in}

\item{file_name}{File name of the data file}

\item{diags}{Diagnosis codes to download}

\item{overwrite_file}{Do you want to overwrite the file if it already exists?}

\item{...}{Not used}
}
\description{
A function to extract NorSySS diagnoses, aggregated
by day and municipality.
}
\examples{
\dontrun{
norsyss_download_aggregated_diagnoses(
  date_from = "2020-01-01",
  date_to = lubridate::today(),
  folder = getwd(),
  diags = list("influenza" = c("R80"))
)
}
}
