% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norsyss_aggregate.R
\name{norsyss_download_new_data}
\alias{norsyss_download_new_data}
\title{norsyss_download_new_data}
\usage{
norsyss_download_new_data(
  confirm = FALSE,
  date_from = "2006-01-01",
  date_to = lubridate::today(),
  ages = c(`0-4` = "0-4", `5-14` = "5-9", `5-14` = "10-14", `15-19` = "15-19", `20-29` =
    "20-29", `30-64` = "30-39", `30-64` = "40-49", `30-64` = "50-59", `30-64` = "60-64",
    `65-69` = "65+", `70-79` = "65+", `80+` = "65+"),
  folder = fs::path(config$folder_sykdomspulsen, "data_raw", "sykdomspuls"),
  overwrite_file = FALSE
)
}
\arguments{
\item{confirm}{a}

\item{date_from}{a}

\item{date_to}{a}

\item{ages}{a}

\item{folder}{a}

\item{overwrite_file}{a}
}
\description{
norsyss_download_new_data
}
