% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norsyss_database.R
\name{norsyss_download_raw}
\alias{norsyss_download_raw}
\title{Download raw data from NorSySS database}
\usage{
norsyss_download_raw(con, date_from, date_to)
}
\arguments{
\item{con}{DB connection}

\item{date_from}{Start date}

\item{date_to}{End date}
}
\description{
Download raw data from NorSySS database
}
